Storage improvements (#341)
* add S3 storage tests, minor improvements

* ensure serial execution of storage tests

* add exceptions around minio api calls, minor changes

+++ nvflare/app_common/storages/s3_storage.py
@@ -13,54 +13,52 @@
 # limitations under the License.
 
 import ast
-import inspect
 import io
-from functools import wraps
+import os
+from pathlib import Path
 from typing import ByteString, List, Tuple
 
 from minio import Minio
 from minio.commonconfig import REPLACE, CopySource
 
 from nvflare.apis.storage import StorageSpec
+from nvflare.apis.utils.format_check import validate_class_methods_args
 
-
-def validate_class_methods_args(cls):
-    for name, method in inspect.getmembers(cls, inspect.isfunction):
-        if name != "__init_subclass__":
-            setattr(cls, name, validate_args(method))
-    return cls
-
-
-def validate_args(method):
-    signature = inspect.signature(method)
-
-    @wraps(method)
-    def wrapper(*args, **kwargs):
-        bound_arguments = signature.bind(*args, **kwargs)
-        for name, value in bound_arguments.arguments.items():
-            annotation = signature.parameters[name].annotation
-            if not (annotation is inspect.Signature.empty or isinstance(value, annotation)):
-                raise TypeError(
-                    "argument '{}' of {} must be {} but got {}".format(name, method, annotation, type(value))
-                )
-        return method(*args, **kwargs)
-
-    return wrapper
+URI_ROOT = os.path.abspath(os.sep)
+_USER_META_KEY = "user_metadata"
 
 
 @validate_class_methods_args
 class S3Storage(StorageSpec):
     def __init__(self, endpoint, access_key, secret_key, secure, bucket_name):
-        self.s3_client = Minio(endpoint=endpoint, access_key=access_key, secret_key=secret_key, secure=secure)
+        """Init S3Storage.
+
+        Uses S3 bucket to persist objects, with absolute paths as object URIs.
+
+        Args:
+            endpoint: hostname of S3 service
+            access_key: access key (username) of S3 service account
+            secret_key: secret key (password) of S3 service account
+            secure: flag for secure (TLS) mode
+            bucket_name: name of S3 bucket
+
+        Raises:
+            RuntimeError: error creating minio S3 client and bucket
+
+        """
+        try:
+            self.s3_client = Minio(endpoint=endpoint, access_key=access_key, secret_key=secret_key, secure=secure)
+            if not self.s3_client.bucket_exists(bucket_name):
+                self.s3_client.make_bucket(bucket_name)
+        except Exception as e:
+            raise RuntimeError("Error creating minio s3 client: {}".format(str(e)))
         self.bucket_name = bucket_name
 
     def _object_exists(self, uri: str):
         try:
             self.s3_client.stat_object(self.bucket_name, uri)
-        except Exception as e:
-            if e.message == "Object does not exist":
-                return False
-            raise e
+        except:
+            return False
         return True
 
     def create_object(self, uri: str, data: ByteString, meta: dict, overwrite_existing: bool = False):
@@ -74,12 +72,13 @@ class S3Storage(StorageSpec):
 
         Returns:
 
-        Raises exception when:
-
-        - invalid URI specification
-        - invalid args
-        - object already exists and overwrite_existing is False
-        - error creating the object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError:
+                - if error creating the object
+                - if object already exists and overwrite_existing is False
+                - if object will be inside prexisiting object
+                - if object will be at a non-empty directory
 
         Examples of URI:
 
@@ -91,9 +90,31 @@ class S3Storage(StorageSpec):
         if self._object_exists(uri) and not overwrite_existing:
             raise RuntimeError("object {} already exists and overwrite_existing is False".format(uri))
 
-        self.s3_client.put_object(
-            self.bucket_name, uri, data=io.BytesIO(data), length=-1, metadata={"user_metadata": meta}, part_size=5242880
-        )
+        path_parts = Path(uri).parts
+        for i in range(2, len(path_parts)):
+            parent_path = str(Path(*path_parts[0:i]))
+            if self._object_exists(parent_path):
+                raise RuntimeError("cannot create object {} inside preexisting object {}".format(uri, parent_path))
+
+        try:
+            dir_is_nonempty = list(self.s3_client.list_objects(self.bucket_name, prefix=uri, include_user_meta=True))
+        except Exception as e:
+            raise RuntimeError("error listing objects: {}".format(str(e)))
+
+        if not self._object_exists(uri) and dir_is_nonempty:
+            raise RuntimeError("cannot create object {} at nonempty directory".format(uri))
+
+        try:
+            self.s3_client.put_object(
+                self.bucket_name,
+                uri,
+                data=io.BytesIO(data),
+                length=-1,
+                metadata={_USER_META_KEY: meta},
+                part_size=5242880,
+            )
+        except Exception as e:
+            raise RuntimeError("error putting object into bucket: {}".format(str(e)))
 
     def update_meta(self, uri: str, meta: dict, replace: bool):
         """Update the meta info of the specified object
@@ -105,11 +126,11 @@ class S3Storage(StorageSpec):
 
         Returns:
 
-        Raises exception when:
-
-        - no such object
-        - invalid args
-        - error updating the object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError:
+                - if object does not exist
+                - if error copying object
 
         """
         if not self._object_exists(uri):
@@ -120,13 +141,16 @@ class S3Storage(StorageSpec):
             prev_meta.update(meta)
             meta = prev_meta
 
-        self.s3_client.copy_object(
-            self.bucket_name,
-            uri,
-            CopySource(self.bucket_name, uri),
-            metadata={"user_metadata": meta},
-            metadata_directive=REPLACE,
-        )
+        try:
+            self.s3_client.copy_object(
+                self.bucket_name,
+                uri,
+                CopySource(self.bucket_name, uri),
+                metadata={_USER_META_KEY: meta},
+                metadata_directive=REPLACE,
+            )
+        except Exception as e:
+            raise RuntimeError("error copying object: {}".format(str(e)))
 
     def update_data(self, uri: str, data: ByteString):
         """Update the data info of the specified object
@@ -137,24 +161,27 @@ class S3Storage(StorageSpec):
 
         Returns:
 
-        Raises exception when:
-
-        - no such object
-        - invalid args
-        - error updating the object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError:
+                - if object does not exist
+                - if error putting object into bucket
 
         """
         if not self._object_exists(uri):
             raise RuntimeError("object {} does not exist".format(uri))
 
-        self.s3_client.put_object(
-            self.bucket_name,
-            uri,
-            data=io.BytesIO(data),
-            length=-1,
-            metadata={"user_metadata": self.get_meta(uri)},
-            part_size=5242880,
-        )
+        try:
+            self.s3_client.put_object(
+                self.bucket_name,
+                uri,
+                data=io.BytesIO(data),
+                length=-1,
+                metadata={_USER_META_KEY: self.get_meta(uri)},
+                part_size=5242880,
+            )
+        except Exception as e:
+            raise RuntimeError("error putting object into bucket: {}".format(str(e)))
 
     def list_objects(self, dir_path: str) -> List[str]:
         """List all objects in the specified path.
@@ -162,15 +189,23 @@ class S3Storage(StorageSpec):
         Args:
             path: the path to the objects
 
-        Returns: list of URIs of objects
+        Returns:
+            list of URIs of objects
+
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError: if error listing objects
 
         """
-        dir_path = dir_path.rstrip("/") + "/"
-        return [
-            "/" + obj._object_name
-            for obj in list(self.s3_client.list_objects(self.bucket_name, prefix=dir_path, include_user_meta=True))
-            if obj._metadata
-        ]
+        dir_path = dir_path.rstrip(os.sep) + os.sep
+        try:
+            return [
+                URI_ROOT + obj._object_name
+                for obj in list(self.s3_client.list_objects(self.bucket_name, prefix=dir_path, include_user_meta=True))
+                if obj._metadata
+            ]
+        except Exception as e:
+            raise RuntimeError("error listing objects: {}".format(str(e)))
 
     def get_meta(self, uri: str) -> dict:
         """Get user defined meta info of the specified object
@@ -178,17 +213,25 @@ class S3Storage(StorageSpec):
         Args:
             uri: URI of the object
 
-        Returns: meta info of the object.
+        Returns:
+            meta info of the object.
 
-        Raises exception when:
-
-        - no such object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError:
+                - if object does not exist
+                - if error accessing object
 
         """
         if not self._object_exists(uri):
             raise RuntimeError("object {} does not exist".format(uri))
 
-        return ast.literal_eval(self.s3_client.stat_object(self.bucket_name, uri)._metadata["x-amz-meta-user_metadata"])
+        try:
+            return ast.literal_eval(
+                self.s3_client.stat_object(self.bucket_name, uri)._metadata["x-amz-meta-{}".format(_USER_META_KEY)]
+            )
+        except Exception as e:
+            raise RuntimeError("error accessing object: {}".format(str(e)))
 
     def get_full_meta(self, uri: str) -> dict:
         """Get full meta info of the specified object
@@ -196,17 +239,23 @@ class S3Storage(StorageSpec):
         Args:
             uri: URI of the object
 
-        Returns: meta info of the object.
+        Returns:
+            meta info of the object.
 
-        Raises exception when:
-
-        - no such object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError:
+                - if object does not exist
+                - if error accessing object
 
         """
         if not self._object_exists(uri):
             raise RuntimeError("object {} does not exist".format(uri))
 
-        return self.s3_client.stat_object(self.bucket_name, uri)
+        try:
+            return self.s3_client.stat_object(self.bucket_name, uri)
+        except Exception as e:
+            raise RuntimeError("error accessing object: {}".format(str(e)))
 
     def get_data(self, uri: str) -> bytes:
         """Get data of the specified object
@@ -214,17 +263,23 @@ class S3Storage(StorageSpec):
         Args:
             uri: URI of the object
 
-        Returns: data of the object.
+        Returns:
+            data of the object.
 
-        Raises exception when:
-
-        - no such object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError:
+                - if object does not exist
+                - if error accessing object
 
         """
         if not self._object_exists(uri):
             raise RuntimeError("object {} does not exist".format(uri))
 
-        return self.s3_client.get_object(self.bucket_name, uri).data
+        try:
+            return self.s3_client.get_object(self.bucket_name, uri).data
+        except Exception as e:
+            raise RuntimeError("error accessing object: {}".format(str(e)))
 
     def get_detail(self, uri: str) -> Tuple[dict, bytes]:
         """Get both data and meta of the specified object
@@ -232,11 +287,12 @@ class S3Storage(StorageSpec):
         Args:
             uri: URI of the object
 
-        Returns: meta info and data of the object.
+        Returns:
+            meta info and data of the object.
 
-        Raises exception when:
-
-        - no such object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError: if object does not exist
 
         """
         if not self._object_exists(uri):
@@ -252,12 +308,17 @@ class S3Storage(StorageSpec):
 
         Returns:
 
-        Raises exception when:
-
-        - no such object
+        Raises:
+            TypeError: if invalid argument types
+            RuntimeError:
+                - if object does not exist
+                - if error removing object
 
         """
         if not self._object_exists(uri):
             raise RuntimeError("object {} does not exist".format(uri))
 
-        self.s3_client.remove_object(self.bucket_name, uri)
+        try:
+            self.s3_client.remove_object(self.bucket_name, uri)
+        except Exception as e:
+            raise RuntimeError("error removing object: {}".format(str(e)))
